Write a Java Program to Count sum and average of all numbers entered by user using infinite loop. If the number is divisible by 5, the loop should continue ( i.e it should not be added in sum). If the number is 999 the loop should be terminated and it will display sum and average both.
Write a Java Program to Count sum and average of all numbers entered by user
using infinite loop. If the number is divisible by 5, the loop should continue ( i.e it should
not be added in sum). If the number is 999 the loop should be terminated and it will
display sum and average both.